git-svn-id: https://libmpplas.svn.sourceforge.net/svnroot/libmpplas/libmpplas/trunk@711 ea26bf2b-611e-0410-8ca6-ad9ecb09ac58

+++ src/headers/GCD.h
@@ -190,13 +190,12 @@ namespace mpplas{
         }
         else{
 #undef min
-          const int real_cs = std::min( chunk_size, (int)listSize - i );
+          const int real_cs = std::min( chunk_size, listSize - i );
           for( int j = 0; (j < real_cs-2) && cont; j++){
             d = gcd(d, nums[i+2+j]);
             if( d == T::getMultIdentity() ){
 #pragma omp atomic
               cont--;
-
             }
           }
         }
